Development

+++ src/main/java/com/novick/customers/landing/repositories/HeaderLinksRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface HeaderLinksRepository extends JpaRepository<HeaderLinks, Long> {
+public interface HeaderLinksRepository extends JpaRepository<HeaderLinks, Integer> {
 }
